refactor: Update ecommerce-main-app routes. Move NotFound page to ecommerce-main-app. Apply laze importing to Shopping in ecommerce-main-app.
